Refactored Song metadata out of Song.java, updated HostActivity's calls to access Song data, and updated the renameAPK bash script
