+++ utilities/renameAPK.sh
@@ -1,1 +1,1 @@
-mv fanClub.apk $(git log --pretty=format:'%h' -n 1).apk
+mv ../signed-apks/fanClub.apk ../signed-apks/$(git log --pretty=format:'%h' -n 1).apk
